Increaze size of outcome struct
The outcome nonces only occupy 24 bytes. But to compute them,
we need a pointer to 32 bytes. Instead of allocating the 32 bytes
in the outcome functions and copying the nonces into the outcome
struct, this commit increases the size of the outcome struct.
This way, the result of the nonce compuation can be written into
the outcome struct directly.